+++ tools/assembler/main.cpp
@@ -1,0 +1,9 @@
+#include <cstdio>
+
+int main(int argc, char** argv) {
+    (void)argc;
+    (void)argv;
+
+    std::puts("stackvm-asm: not implemented yet");
+    return 0;
+}